requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-simple-history==3.5.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
